Make gen_build_version.py resilient to a failing git rev-parse
It was noticed that some build processes did not checkout Impala and
instead built it from a tarball. This would cause our gen_build_version
script to write a blank version info everytime to the version.info file.

This patch takes care of the case where if there is an already existing
version.info file and we cannot get the git rev-parse output, we use
the old file instead. Blank version info is written only when we don't
have an old version.info file and we cannot do a git rev-parse.

Reviewed-on: http://gerrit.cloudera.org:8080/4411
Tested-by: Sailesh Mukil <sailesh@cloudera.com>

+++ bin/gen_build_version.py
@@ -23,6 +23,7 @@
 import os
 import time;
 import filecmp
+from subprocess import PIPE, call
 from commands import getstatusoutput
 from time import localtime, strftime
 
@@ -31,17 +32,34 @@ SAVE_VERSION_SCRIPT = os.path.join(IMPALA_HOME, 'bin/save-version.sh')
 VERSION_FILE_NAME = os.path.join(IMPALA_HOME, 'bin/version.info')
 VERSION_CC_FILE_NAME = os.path.join(IMPALA_HOME, 'be/src/common/version.cc')
 
-# Remove existing version files only if they exist.
-# TODO: Might be useful to make a common utility function remove_if_clean.
-if os.path.isfile(VERSION_FILE_NAME):
-  print 'Removing existing file: %s' % (VERSION_FILE_NAME)
-  os.remove(VERSION_FILE_NAME)
-if os.path.isfile(VERSION_CC_FILE_NAME):
-  print 'Removing existing file: %s' % (VERSION_CC_FILE_NAME)
-  os.remove(VERSION_CC_FILE_NAME)
+# Redirecting stdout and stderr to os.devnull as we don't want unnecessary output.
+devnull = open(os.devnull, 'w')
+try:
+  can_obtain_git_hash = \
+      call(['git', 'rev-parse', 'HEAD'], stdout=devnull, stderr=devnull) == 0
+finally:
+  devnull.close()
 
-# Generate a new version file.
-os.system(SAVE_VERSION_SCRIPT)
+version_file_exists = os.path.isfile(VERSION_FILE_NAME)
+
+# If we have a version file and cannot obtain a git hash, skip generating a new
+# version file.
+if version_file_exists and not can_obtain_git_hash:
+  print "Cannot obtain git hash, using existing version file."
+else:
+  # Remove existing version files only if they exist.
+  # TODO: Might be useful to make a common utility function remove_if_clean.
+  if version_file_exists:
+    print 'Removing existing file: %s' % (VERSION_FILE_NAME)
+    os.remove(VERSION_FILE_NAME)
+  if os.path.isfile(VERSION_CC_FILE_NAME):
+    print 'Removing existing file: %s' % (VERSION_CC_FILE_NAME)
+    os.remove(VERSION_CC_FILE_NAME)
+
+  # SAVE_VERSION_SCRIPT will generate a dummy version.info file if we cannot obtain the
+  # git hash.
+  # Generate a new version file.
+  os.system(SAVE_VERSION_SCRIPT)
 
 # version.info file has the format:
 # VERSION: <version>
@@ -56,9 +74,6 @@ finally:
   version_file.close()
 
 print '\n'.join([version, git_hash, build_time])
-
-# construct the build time (e.g. Thu, 04 Oct 2012 11:53:17 PST)
-build_time = "%s %s" % (strftime("%a, %d %b %Y %H:%M:%S", localtime()), time.tzname[0])
 
 file_contents = """
 //
